Implemented basic locks
Implemented a version of locks that passes the kernel tests. The thread sleeps if the lock can not be acquired so there is a risk of starvation and the locks are not the most fair. However, given the project constraints and the vague atomicity of OS161. This is the best implementation I could currently get to work.